+++ py/desitarget/mtl.py
@@ -41,7 +41,7 @@ mtldatamodel = np.array([], dtype=[
     ('SUBPRIORITY', '>f8'), ('OBSCONDITIONS', 'i4'),
     ('PRIORITY_INIT', '>i8'), ('NUMOBS_INIT', '>i8'), ('PRIORITY', '>i8'),
     ('NUMOBS', '>i8'), ('NUMOBS_MORE', '>i8'), ('Z', '>f8'), ('ZWARN', '>i8'),
-    ('TIMESTAMP', 'S19'), ('VERSION', 'S14'), ('TARGET_STATE', 'S16'),
+    ('TIMESTAMP', 'U19'), ('VERSION', 'U14'), ('TARGET_STATE', 'U16'),
     ('ZTILEID', '>i4')
     ])
 
@@ -51,8 +51,8 @@ zcatdatamodel = np.array([], dtype=[
     ])
 
 mtltilefiledm = np.array([], dtype=[
-    ('TILEID', '>i4'), ('TIMESTAMP', 'S19'),
-    ('VERSION', 'S14'), ('PROGRAM', 'S6')
+    ('TILEID', '>i4'), ('TIMESTAMP', 'U19'),
+    ('VERSION', 'U14'), ('PROGRAM', 'U6'), ('ZDATE', 'U8')
     ])
 
 
@@ -131,33 +131,6 @@ def get_zcat_dir(zcatdir=None):
     return zcatdir
 
 
-def get_tile_full_path(tilefn=None):
-    """Convenience function to grab the $TILE_FN environment variable.
-
-    Parameters
-    ----------
-    tilefn : :class:`str`, optional, defaults to $TILE_FN
-        If `tilefn` is passed, it is returned from this function. If it's
-        not passed, the $TILE_FN environment variable is returned.
-
-    Returns
-    -------
-    :class:`str`
-        If `tilefn` is passed, it is returned from this function. If it's
-        not passed, the filename stored in the $TILE_FN environment
-        variable is returned.
-    """
-    if tilefn is None:
-        tilefn = os.environ.get('TILE_FN')
-        # ADM check that the $TILE_FN environment variable is set.
-        if tilefn is None:
-            msg = "Pass tilefn or set $TILE_FN environment variable!"
-            log.critical(msg)
-            raise ValueError(msg)
-
-    return tilefn
-
-
 def get_mtl_tile_file_name():
     """Convenience function to grab the name of the MTL tile file.
 
@@ -167,6 +140,19 @@ def get_mtl_tile_file_name():
         The name of the MTL tile file.
     """
     fn = "mtl-done-tiles.ecsv"
+
+    return fn
+
+
+def get_ztile_file_name():
+    """Convenience function to grab the name of the ZTILE file.
+
+    Returns
+    -------
+    :class:`str`
+        The name of the ZTILE file.
+    """
+    fn = "tiles.csv"
 
     return fn
 
@@ -353,7 +339,7 @@ def make_mtl(targets, obscon, zcat=None, scnd=None,
                                     dtype=mtldatamodel["SCND_TARGET"].dtype)
 
     # ADM initialize columns to avoid zero-length/missing/format errors.
-    zcols = ["NUMOBS_MORE", "NUMOBS", "Z", "ZWARN"]
+    zcols = ["NUMOBS_MORE", "NUMOBS", "Z", "ZWARN", "ZTILEID"]
     for col in zcols + ["TARGET_STATE", "TIMESTAMP", "VERSION"]:
         mtl[col] = np.empty(len(mtl), dtype=mtldatamodel[col].dtype)
 
@@ -374,7 +360,7 @@ def make_mtl(targets, obscon, zcat=None, scnd=None,
         mtl[col][zmatcher] = ztargets[col]
     # ADM also add the ZTILEID column, if passed, otherwise we're likely
     # ADM to be working with non-ledger-based mocks and can let it slide.
-    if "ZTILEID" in ztargets:
+    if "ZTILEID" in ztargets.dtype.names:
         mtl["ZTILEID"][zmatcher] = ztargets["ZTILEID"]
     else:
         mtl["ZTILEID"] = -1
@@ -406,8 +392,8 @@ def make_mtl(targets, obscon, zcat=None, scnd=None,
     return mtl
 
 
-def make_ledger_in_hp(targets, outdirname, nside, pixlist,
-                      obscon="DARK", indirname=None, verbose=True):
+def make_ledger_in_hp(targets, outdirname, nside, pixlist, obscon="DARK",
+                      indirname=None, verbose=True, scnd=False):
     """
     Make an initial MTL ledger file for targets in a set of HEALPixels.
 
@@ -432,6 +418,8 @@ def make_ledger_in_hp(targets, outdirname, nside, pixlist,
         of the output MTL files.
     verbose : :class:`bool`, optional, defaults to ``True``
         If ``True`` then log target and file information.
+    scnd : :class:`bool`, defaults to ``False``
+        If ``True`` then this is a ledger of secondary targets.
 
     Returns
     -------
@@ -455,12 +443,14 @@ def make_ledger_in_hp(targets, outdirname, nside, pixlist,
     for pix in pixlist:
         inpix = mtlpix == pix
         ecsv = get_mtl_ledger_format() == "ecsv"
-        nt, fn = io.write_mtl(
-            outdirname, mtl[inpix].as_array(), indir=indirname, ecsv=ecsv,
-            survey=survey, obscon=obscon, nsidefile=nside, hpxlist=pix)
-        if verbose:
-            log.info('{} targets written to {}...t={:.1f}s'.format(
-                nt, fn, time()-t0))
+        if np.any(inpix):
+            nt, fn = io.write_mtl(
+                outdirname, mtl[inpix].as_array(), indir=indirname, ecsv=ecsv,
+                survey=survey, obscon=obscon, nsidefile=nside, hpxlist=pix,
+                scnd=scnd)
+            if verbose:
+                log.info('{} targets written to {}...t={:.1f}s'.format(
+                    nt, fn, time()-t0))
 
     return
 
@@ -512,13 +502,23 @@ def make_ledger(hpdirname, outdirname, pixlist=None, obscon="DARK", numproc=1):
         log.critical(msg)
         raise ValueError(msg)
 
+    # ADM check whether this is a file of standalone secondary targets.
+    scnd = False
+    if hdr["EXTNAME"] == 'SCND_TARGETS':
+        scnd = True
+
     # ADM the MTL datamodel must reflect the target flavor (SV, etc.).
     mtldm = switch_main_cmx_or_sv(mtldatamodel, np.array([], dt))
     # ADM speed-up by only reading the necessary columns.
     cols = list(set(mtldm.dtype.names).intersection(dt.names))
 
     # ADM optimal nside for reading in the targeting files.
-    nside = hdr["FILENSID"]
+    if "FILENSID" in hdr:
+        nside = hdr["FILENSID"]
+    else:
+        # ADM if a file was passed instead of a HEALPixel-split directory
+        # ADM use nside=4 as 196 pixels often balances well across CPUs.
+        nside = 4
     npixels = hp.nside2npix(nside)
     pixels = np.arange(npixels)
 
@@ -549,10 +549,20 @@ def make_ledger(hpdirname, outdirname, pixlist=None, obscon="DARK", numproc=1):
         targs = io.read_targets_in_hp(hpdirname, nside, pixnum, columns=cols)
         if len(targs) == 0:
             return
+        # ADM the secondary targeting files don't include the BGS_TARGET
+        # ADM and MWS_TARGET columns, which are needed for MTL.
+        neededcols, _, _ = main_cmx_or_sv(targs)
+        misscols = set(neededcols) - set(cols)
+        if len(misscols) > 0:
+            # ADM the data type for the DESI_TARGET column.
+            dtdt = mtldatamodel["DESI_TARGET"].dtype
+            zerod = [np.zeros(len(targs)), np.zeros(len(targs))]
+            targs = rfn.append_fields(
+                targs, misscols, data=zerod, dtypes=[dtdt, dtdt], usemask=False)
         # ADM write MTLs for the targs split over HEALPixels in pixlist.
         return make_ledger_in_hp(
-            targs, outdirname, mtlnside, pix,
-            obscon=obscon, indirname=hpdirname, verbose=False)
+            targs, outdirname, mtlnside, pix, obscon=obscon,
+            indirname=hpdirname, verbose=False, scnd=scnd)
 
     # ADM this is just to count pixels in _update_status.
     npix = np.ones((), dtype='i8')
@@ -776,22 +786,23 @@ def inflate_ledger(mtl, hpdirname, columns=None, header=False, strictcols=False,
     return done
 
 
-def tiles_to_be_processed(zcatdir, mtltilefn, tilefn, obscon):
-    """Execute full MTL loop, including reading files, updating ledgers.
+def tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey):
+    """Find tiles that are "done" but aren't yet in the MTL tile record.
 
     Parameters
     ----------
     zcatdir : :class:`str`
-        Full path to the directory that hosts the redshift catalogs.
+        Full path to the "daily" directory that hosts redshift catalogs.
     mtltilefn : :class:`str`
         Full path to the file of tiles that have been processed by MTL.
-    tilefn : :class:`str`, optional, defaults to ``None``
-        Full path to the name of the tile file. This file is used to link
-        TILEIDs to observing conditions.
     obscon : :class:`str`
         A string matching ONE obscondition in the desitarget bitmask yaml
         file (i.e. in `desitarget.targetmask.obsconditions`), e.g. "DARK"
         Governs how priorities are set when merging targets.
+    survey : :class:`str`, optional, defaults to "main"
+        Used to look up the correct ledger, in combination with `obscon`.
+        Options are ``'main'`` and ``'svX``' (where X is 1, 2, 3 etc.)
+        for the main survey and different iterations of SV, respectively.
 
     Returns
     -------
@@ -799,18 +810,22 @@ def tiles_to_be_processed(zcatdir, mtltilefn, tilefn, obscon):
         An array of tiles that have not yet been processed and written to
         the mtl tile file.
     """
-    # ADM read in the tile file.
-    tilelookup = Table.read(tilefn)
+    # ADM read in the ZTILE file.
+    ztilefn = os.path.join(zcatdir, get_ztile_file_name())
+    tilelookup = Table.read(ztilefn)
+
+    # ADM the ZDONE column is a string, convert to a Boolean.
+    zdone = np.array(["true" in row for row in tilelookup["ZDONE"]])
 
     # ADM determine the tiles to be processed.
-    alltiles = []
-    for fn in glob(os.path.join(zcatdir, "*")):
-        # ADM sometimes there are weird tile names like "temp".
-        try:
-            tileid = int(os.path.basename(fn))
-        except ValueError:
-            pass
-        alltiles.append(tileid)
+    # ADM must match the correct survey.
+    ii = tilelookup["SURVEY"] == survey
+    # ADM must match the correct OBSCON, could be lower- or upper-case.
+    ii &= ((tilelookup["FAPRGRM"] == obscon.lower()) |
+           (tilelookup["FAPRGRM"] == obscon.upper()))
+    # ADM redshift processing must be complete.
+    ii &= zdone
+    alltiles = tilelookup["TILEID"][ii]
 
     # ADM read in the tile file, guarding against it not having being
     # ADM created yet.
@@ -820,12 +835,9 @@ def tiles_to_be_processed(zcatdir, mtltilefn, tilefn, obscon):
 
     # ADM extract the updated tiles.
     if donetiles is None:
-        tilesallcon = np.array(alltiles)
+        tileids = np.array(alltiles)
     else:
-        tilesallcon = np.array(list(set(alltiles) - set(donetiles["TILEID"])))
-
-    obstiles = tilelookup[tilelookup["PROGRAM"] == obscon]["TILEID"]
-    tileids = list(set(obstiles).intersection(set(tilesallcon)))
+        tileids = np.array(list(set(alltiles) - set(donetiles["TILEID"])))
 
     # ADM initialize the output array and add the tiles.
     donetiles = np.zeros(len(tileids), dtype=mtltilefiledm.dtype)
@@ -836,6 +848,8 @@ def tiles_to_be_processed(zcatdir, mtltilefn, tilefn, obscon):
     donetiles["VERSION"] = dt_version
     # ADM add the program/obscon.
     donetiles["PROGRAM"] = obscon
+    # ADM add a placeholder for the YYYYMMDD of tile redshifts.
+    donetiles["ZDATE"] = ""
 
     return donetiles
 
@@ -846,7 +860,7 @@ def make_zcat_rr_backstop(zcatdir, tiles):
     Parameters
     ----------
     zcatdir : :class:`str`
-        Full path to the directory that hosts the redshift catalogs.
+        Full path to the "daily" directory that hosts redshift catalogs.
     tiles : :class:`~numpy.array`
         List of TILEIDs in `zcatdir` from which to construct the `zcat`.
 
@@ -855,20 +869,40 @@ def make_zcat_rr_backstop(zcatdir, tiles):
     :class:`~astropy.table.Table`
         A zcat in the official format (`zcatdatamodel`) compiled from
         the `tiles` in `zcatdir`.
+    :class:`list`
+        The YEARMMDD timestamp that corresponds to each tile in the
+        input `tiles`.
 
     Notes
     -----
     - How the `zcat` is constructed could certainly change once we have
       the final schema in place.
     """
+    # ADM the root directory in the data model.
+    rootdir = os.path.join(zcatdir, "tiles", "cumulative")
+
     # ADM for each tile, read in the spectroscopic and targeting info.
     allzs = []
     allfms = []
+    ymds = []
     for tile in tiles:
-        zbestfns = os.path.join(zcatdir, "{}".format(tile), "*", "zbest*")
-        for zbestfn in glob(zbestfns):
-            allzs.append(fitsio.read(zbestfn, "ZBEST"))
-            fm = fitsio.read(zbestfn, "FIBERMAP")
+        # ADM build the correct directory structure.
+        tiledir = os.path.join(rootdir, str(tile))
+        ymd = os.listdir(tiledir)
+        # ADM there should only be one date in the cumulative directory.
+        if len(ymd) != 1:
+            msg = "expected 1 date in {} but found {}".format(tiledir, len(ymd))
+            log.critical(msg)
+            raise OSError(msg)
+        ymdir = os.path.join(tiledir, ymd[0])
+        # ADM record the YYYYMMDD string for this tile.
+        ymds.append(ymd[0])
+        zbestfns = glob(os.path.join(ymdir, "zbest*"))
+        for zbestfn in zbestfns:
+            zz = fitsio.read(zbestfn, "ZBEST")
+            allzs.append(zz)
+            # ADM only read in the first set of exposures.
+            fm = fitsio.read(zbestfn, "FIBERMAP", rows=np.arange(len(zz)))
             allfms.append(fm)
             # ADM check the correct TILEID was written in the fibermap.
             if set(fm["TILEID"]) != set([tile]):
@@ -877,6 +911,10 @@ def make_zcat_rr_backstop(zcatdir, tiles):
                 raise ValueError(msg)
     zs = np.concatenate(allzs)
     fms = np.concatenate(allfms)
+
+    # ADM remove -ve TARGETIDs which should correspond to bad fibers.
+    zs = zs[zs["TARGETID"] >= 0]
+    fms = fms[fms["TARGETID"] >= 0]
 
     # ADM currently, the spectroscopic files aren't coadds, so aren't
     # ADM unique. We therefore need to look up (any) coordinates for
@@ -893,11 +931,11 @@ def make_zcat_rr_backstop(zcatdir, tiles):
     for col in set(zcat.dtype.names) - set(['RA', 'DEC', 'NUMOBS', 'ZTILEID']):
         zcat[col] = zs[col]
 
-    return zcat
+    return zcat, ymds
 
 
 def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
-                tilefn=None, numobs_from_ledger=True):
+                numobs_from_ledger=True):
     """Execute full MTL loop, including reading files, updating ledgers.
 
     Parameters
@@ -907,21 +945,17 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
         file (i.e. in `desitarget.targetmask.obsconditions`), e.g. "DARK"
         Governs how priorities are set when merging targets.
     survey : :class:`str`, optional, defaults to "main"
-        Used to look up the correct filename for `obscon`. Options are
-        ``'main'`` and ``'svX``' (where X is 1, 2, 3 etc.) for the main
-        survey and different iterations of SV, respectively.
+        Used to look up the correct ledger, in combination with `obscon`.
+        Options are ``'main'`` and ``'svX``' (where X is 1, 2, 3 etc.)
+        for the main survey and different iterations of SV, respectively.
     zcatdir : :class:`str`, optional, defaults to ``None``
-        Full path to the directory that hosts the redshift catalogs.
+        Full path to the "daily" directory that hosts redshift catalogs.
         If this is ``None``, look up the redshift catalog directory from
         the $ZCAT_DIR environment variable.
     mtldir : :class:`str`, optional, defaults to ``None``
         Full path to the directory that hosts the MTL ledgers and the MTL
-        tile file. If this ``None``, look up the MTL directory from the
+        tile file. If ``None``, then look up the MTL directory from the
         $MTL_DIR environment variable.
-    tilefn : :class:`str`, optional, defaults to ``None``
-        Full path to the name of the tile file. This file is used to link
-        TILEIDs to observing conditions. If passed as ``None``, looked up
-        from the $TILE_FN environment variable.
     numobs_from_ledger : :class:`bool`, optional, defaults to ``True``
         If ``True`` then inherit the number of observations so far from
         the ledger rather than expecting it to have a reasonable value
@@ -934,24 +968,21 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     :class:`str`
         The name of the MTL tile file that was updated.
     :class:`str`
-        The name of the tile file that was used to link TILEIDs to
-        observing conditions.
+        The name of the ZTILE file that was used to link TILEIDs to
+        observing conditions and to determine if tiles were "done".
     :class:`~numpy.array`
         Information for the tiles that were processed.
 
     Notes
     -----
     - Assumes all of the relevant ledgers have already been made by,
-      e.g., `make_ledger()`.
+      e.g., :func:`~desitarget.mtl.make_ledger()`.
     """
     # ADM first grab all of the relevant files.
     # ADM grab the MTL directory (in case we're relying on $MTL_DIR).
     mtldir = get_mtl_dir(mtldir)
     # ADM construct the full path to the mtl tile file.
     mtltilefn = os.path.join(mtldir, get_mtl_tile_file_name())
-
-    # ADM grab the TILE filename (in case we're relying on $TILE_FN).
-    tilefn = get_tile_full_path(tilefn)
 
     # ADM construct the relevant sub-directory for this survey and
     # ADM set of observing conditions..
@@ -960,17 +991,21 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
                                      survey=survey, obscon=obscon, ender=form)
     # ADM grab the zcat directory (in case we're relying on $ZCAT_DIR).
     zcatdir = get_zcat_dir(zcatdir)
+    # ADM And contruct the associated ZTILE filename.
+    ztilefn = os.path.join(zcatdir, get_ztile_file_name())
 
     # ADM grab an array of tiles that are yet to be processed.
-    tiles = tiles_to_be_processed(zcatdir, mtltilefn, tilefn, obscon)
+    tiles = tiles_to_be_processed(zcatdir, mtltilefn, obscon, survey)
 
     # ADM stop if there are no tiles to process.
     if len(tiles) == 0:
-        return hpdirname, mtltilefn, tilefn, tiles
+        return hpdirname, mtltilefn, ztilefn, tiles
 
     # ADM create the zcat: This will likely change, but for now let's
     # ADM just use redrock in the SV1-era format.
-    zcat = make_zcat_rr_backstop(zcatdir, tiles["TILEID"])
+    zcat, ymd = make_zcat_rr_backstop(zcatdir, tiles["TILEID"])
+    # ADM update the tiles table with the YYYYMMDD.
+    tiles["ZDATE"] = ymd
     # ADM insist that for an MTL loop with real observations, the zcat
     # ADM must conform to the data model. In particular, it must include
     # ADM ZTILEID, which may not be needed for non-ledger simulations.
@@ -987,4 +1022,4 @@ def loop_ledger(obscon, survey='main', zcatdir=None, mtldir=None,
     # ADM write the processed tiles to the MTL tile file.
     io.write_mtl_tile_file(mtltilefn, tiles)
 
-    return hpdirname, mtltilefn, tilefn, tiles
+    return hpdirname, mtltilefn, ztilefn, tiles