+++ doc/changes.rst
@@ -5,7 +5,16 @@ desitarget Change Log
 0.53.1 (unreleased)
 -------------------
 
-* Change required minimum r_value for QSO RF selection from 17.5 to 16.5 [`PR #692`]
+* Update MTL loop for new data model [`PR #693`_]. Includes:
+    * Functionality to make initial ledgers for secondary targets.
+    * Use the ``ZTILE`` file to look-up redshift "done-ness" (``zdone``)
+        * instead of relying on a "done" directory.
+    * Also derive the survey and program/obscon from the ``ZTILE`` file.
+    * Switch to reading the ``zbest`` files in the cumulative directory.
+    * Remove -ve ``TARGETIDs`` before matching.
+    * Add the date that redshifts were extracted to the mtl tiles file.
+    * Add the root redshift directory (``ZCAT_DIR``) to the manifest.
+* Change minimum r_value for QSO RF from 17.5 to 16.5 [`PR #692`].
 * Some clean-up for the 1% Survey [`PR #691`_]. Includes:
     * Don't allow ``BGS_FAINT`` targets to be observed in ``DARK``.
     * Warn about primary targets that might be too bright.
@@ -25,6 +34,7 @@ desitarget Change Log
 .. _`PR #690`: https://github.com/desihub/desitarget/pull/690
 .. _`PR #691`: https://github.com/desihub/desitarget/pull/691
 .. _`PR #692`: https://github.com/desihub/desitarget/pull/692
+.. _`PR #693`: https://github.com/desihub/desitarget/pull/693
 
 0.53.0 (2021-03-18)
 -------------------